--- a/ClosedXML/Excel/XLWorkbook_Load.cs
+++ b/ClosedXML/Excel/XLWorkbook_Load.cs
@@ -1379,14 +1379,15 @@ namespace ClosedXML.Excel
         {
             Int32 styleIndex = cell.StyleIndex != null ? Int32.Parse(cell.StyleIndex.InnerText) : 0;
 
-            String cellReference = cell.CellReference == null
-                                       ? XLHelper.GetColumnLetterFromNumber(++lastCell) + rowIndex
-                                       : cell.CellReference.Value;
-            var xlCell = ws.CellFast(cellReference);
+            var cellAddress = cell.CellReference == null
+                                       ? new XLAddress(ws, ++lastCell, rowIndex, false, false)
+                                       : XLAddress.Create(ws, cell.CellReference.Value);
+
+            var xlCell = ws.Cell(in cellAddress);
 
             if (styleList.ContainsKey(styleIndex))
             {
-                xlCell.Style = styleList[styleIndex];
+                xlCell.InnerStyle = styleList[styleIndex];
             }
             else
             {
@@ -1566,14 +1567,14 @@ namespace ClosedXML.Excel
                     if (s.NumberingFormats != null &&
                         s.NumberingFormats.Any(nf => ((NumberingFormat)nf).NumberFormatId.Value == numberFormatId))
                     {
-                        xlCell.Style.NumberFormat.Format =
+                        xlCell.InnerStyle.NumberFormat.Format =
                             ((NumberingFormat)s.NumberingFormats
                                                 .First(
                                                     nf => ((NumberingFormat)nf).NumberFormatId.Value == numberFormatId)
                             ).FormatCode.Value;
                     }
                     else
-                        xlCell.Style.NumberFormat.NumberFormatId = Int32.Parse(numberFormatId);
+                        xlCell.InnerStyle.NumberFormat.NumberFormatId = Int32.Parse(numberFormatId);
                 }
             }