Here are a few changes that improve the performance of the large file loading. On my test example, the load time decreased from 13 sec to 8.5 sec in Debug mode and from 10 sec to 6.5 sec in Release mode.

- [ ] C# Code Review: @csreviewer
- [ ] Test Automation Review: @csreviewer